test(#83): fix tests

+++ packages/backend/src/api/recipe.js
@@ -57,6 +57,11 @@ route.put('/', async (req, res) => {
 route.delete('/', async (req, res) => {
   try {
     const { name, id } = req.query;
+
+    if (!id) {
+      return res.status(400).json({ message: '\'id\' is missing from query parameter' });
+    }
+
     await service.deleteRecipe(name, id);
     return res.status(200).json({ message: `${name} deleted!` });
   } catch (e) {
